refactor(llm_tool): extract common logic into functions and apply minor fixes
